refact/front/CreateTripBtn.jsx: moves to src/components/buttons/CreateTripBtn.jsx
